Separate unit tests into a separate file
